Demo - Database error V2

+++ demo/src/main/java/com/spa/demo/backend/ServicesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServicesRepository extends JpaRepository<Services, Integer> {
+public interface ServicesRepository extends JpaRepository<Services, Long> {
 
 }
